requirements.txt
django>=4.2.3,<4.2.99
djangorestframework>=3.14.0,<3.14.99
psycopg>=3.1.9,<3.1.99
Pillow>=10.0.0,<10.0.99
whitenoise>=6.5.0,<6.5.99
drf-spectacular>=0.26.3,<0.26.99
factory-boy>=3.2.1,<3.2.99
django-extensions>=3.2.0,<3.2.99
flake8>=6.0.0,<6.0.99
isort>=5.12.0,<5.12.99
black==23.7.0
celery>=5.3.1,<5.3.99
coverage>=7.2.7,<7.2.99
django-otp>=1.2.2,<1.2.99
qrcode==7.4.2
django-redis>=5.3.0,<5.3.99
PyJWT>=2.7.0,<2.7.99
djangorestframework-simplejwt>=5.2.2,<5.2.99
django-cors-headers>=4.2.0,<4.2.99
django-debug-toolbar>=4.1.0,<4.1.99
sentry-sdk>=1.28.0,<1.28.99
